moved the js files
i moved jx files into the js folder and updated the import for my js files
